修改安全验证相关代码

+++ yougou-mall-security/src/main/java/per/ccm/ygmall/security/vo/TokenVO.java
@@ -15,5 +15,5 @@ public class TokenVO extends BaseVO {
     private String accessToken;
 
     @Schema(description = "过期时间")
-    private Integer expiredIn;
+    private Long expiredIn;
 }
